services/vision_ui/app/requirements.txt
Pillow==9.5.0
Requests==2.30.0
streamlit==1.26.0
streamlit_lottie==0.0.3
streamlit_timeline==0.0.2
walrus
opencv-python==4.5.4.60
pandas
